gh-2220 Restructured the import key flows
