+++ Multisig/UI/Settings/Passcode/PasscodeViewController.swift
@@ -141,305 +141,3 @@ class PasscodeViewController: UIViewController, UITextFieldDelegate {
         }
     }
 }
-
-
-class CreatePasscodeViewController: PasscodeViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        navigationItem.title = "Create Passcode"
-    }
-
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        navigationController?.setNavigationBarHidden(false, animated: false)
-    }
-
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        Tracker.trackEvent(.createPasscode)
-    }
-
-    override func willChangeText(_ text: String) {
-        super.willChangeText(text)
-        guard text.count == passcodeLength else { return }
-
-        // repeat the same passcode
-        let repeatVC = RepeatPasscodeViewController(passcode: text)
-        navigationController?.pushViewController(repeatVC, animated: true)
-
-        // after that, enable biometry
-
-        repeatVC.completion = { [weak self, weak repeatVC] in
-            //  when repeated successfully
-
-            // if device does not support biometrics, finish right away
-            guard App.shared.auth.isBiometricsSupported else {
-                self?.completion()
-                return
-            }
-
-            //   if device supports it, ask if to enable biometry
-            let shouldEnableVC = UIAlertController(
-                title: "Activate Biometry?",
-                message: "Would you like to enable login with biometrics?",
-                preferredStyle: .alert)
-
-            //      if yes, ask to authenticate with biometry
-            shouldEnableVC.addAction(UIAlertAction(title: "Enable", style: .default, handler: { [weak self] _ in
-
-                App.shared.auth.activateBiometrics { _ in
-                    // in any resulting case, finish.
-                    self?.completion()
-                }
-
-            }))
-
-            //      if no, finish right away
-            shouldEnableVC.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: { [weak self] _ in
-                self?.completion()
-            }))
-
-            repeatVC?.present(shouldEnableVC, animated: true, completion: nil)
-        }
-
-        repeatVC.skipCompletion = { [weak self] in
-            self?.completion()
-        }
-    }
-
-    override func didTapButton(_ sender: Any) {
-        completion()
-        Tracker.trackEvent(.userPasscodeSkipped)
-    }
-}
-
-class RepeatPasscodeViewController: PasscodeViewController {
-    var passcode: String!
-    var skipCompletion: () -> Void = {}
-
-    convenience init(passcode: String, completionHandler: @escaping () -> Void = {}) {
-        self.init(namedClass: PasscodeViewController.self)
-        self.passcode = passcode
-        self.completion = completionHandler
-    }
-
-    override func didTapButton(_ sender: Any) {
-        Tracker.trackEvent(.userPasscodeSkipped)
-        skipCompletion()
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        navigationItem.title = "Create Passcode"
-        promptLabel.text = "Repeat the 6-digit passcode"
-    }
-
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        Tracker.trackEvent(.repeatPasscode)
-    }
-
-    override func willChangeText(_ text: String) {
-        super.willChangeText(text)
-        errorLabel.isHidden = true
-        if text == passcode {
-            do {
-                try App.shared.auth.createPasscode(plaintextPasscode: text)
-                App.shared.snackbar.show(message: "Passcode created")
-                completion()
-            } catch {
-                showGenericError(description: "Failed to create passcode", error: error)
-            }
-        } else if text.count == passcodeLength {
-            showError("Passcodes don't match")
-        }
-    }
-}
-
-class EnterPasscodeViewController: PasscodeViewController {
-    var passcodeCompletion: (_ success: Bool, _ reset: Bool) -> Void = { _, _ in }
-    var navigationItemTitle = "Enter Passcode"
-    var screenTrackingEvent = TrackingEvent.enterPasscode
-    var showsCloseButton: Bool = true
-    var usesBiometry: Bool = true
-    var warnAfterWrongAttemptCount: Int = 5
-    var wrongAttemptsCount: Int = 0
-
-    convenience init() {
-        self.init(namedClass: PasscodeViewController.self)
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        navigationItem.title = navigationItemTitle
-        promptLabel.text = "Enter your current passcode"
-        button.setText("Forgot your passcode?", .plain)
-        detailLabel.isHidden = true
-
-        if showsCloseButton {
-            navigationItem.leftBarButtonItem = UIBarButtonItem(
-                barButtonSystemItem: .close,
-                target: self,
-                action: #selector(didTapCloseButton))
-        }
-
-        biometryButton.isHidden = !canUseBiometry
-        biometryButton.setImage(App.shared.auth.isFaceID ? UIImage(named: "ic-face-id") : UIImage(named: "ic-touch-id"), for: .normal)
-    }
-
-    private var canUseBiometry: Bool {
-        usesBiometry && App.shared.auth.isBiometryAuthenticationPossible && AppSettings.passcodeOptions.contains(.useBiometry)
-    }
-
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        Tracker.trackEvent(screenTrackingEvent)
-        authenticateWithBiometry()
-    }
-
-    override func willChangeText(_ text: String) {
-        super.willChangeText(text)
-        errorLabel.isHidden = true
-        if text.count == passcodeLength {
-
-            var isCorrect = false
-            do {
-                isCorrect = try App.shared.auth.isPasscodeCorrect(plaintextPasscode: text)
-            } catch {
-                showGenericError(description: "Failed to check passcode", error: error)
-                return
-            }
-
-            if isCorrect {
-                passcodeCompletion(true, false)
-            } else {
-                wrongAttemptsCount += 1
-                if wrongAttemptsCount >= warnAfterWrongAttemptCount {
-                    showError("\(wrongAttemptsCount) failed password attempts. You can reset password via \"Forgot passcode?\" button below.")
-                } else {
-                    showError("Wrong passcode")
-                }
-            }
-        }
-    }
-
-    @objc func didTapCloseButton() {
-        passcodeCompletion(false, false)
-    }
-
-    override func didTapButton(_ sender: Any) {
-        let alertController = UIAlertController(
-            title: "Remove all content",
-            message: "Disabling the passcode will remove all app content. This cannot be undone. Please type in \"Remove\" to continue.",
-            preferredStyle: .alert)
-
-        alertController.addTextField()
-
-        let remove = UIAlertAction(title: "Disable Passcode", style: .destructive) { [unowned self] _ in
-            guard alertController.textFields?.first?.text?.trimmingCharacters(in: .whitespacesAndNewlines) == "Remove" else {
-                return
-            }
-            do {
-                // should be before deleting all data
-                self.passcodeCompletion(false, true)
-                try App.shared.auth.deleteAllData()
-            } catch {
-                showGenericError(description: "Failed to remove passcode", error: error)
-                return
-            }
-        }
-        let cancel = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
-        alertController.addAction(remove)
-        alertController.addAction(cancel)
-        present(alertController, animated: true)
-    }
-
-    override func didTapBiometry(_ sender: Any) {
-        authenticateWithBiometry()
-    }
-
-    private func authenticateWithBiometry() {
-        guard canUseBiometry else { return }
-
-        App.shared.auth.authenticateWithBiometrics { [weak self] result in
-            guard let `self` = self else { return }
-            switch result {
-            case .success:
-                self.passcodeCompletion(true, false)
-
-            case .failure(_):
-                self.biometryButton.isHidden = !self.canUseBiometry
-            }
-        }
-    }
-}
-
-class ChangePasscodeEnterNewViewController: PasscodeViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        navigationItem.title = "Change Passcode"
-        promptLabel.text = "Create a new 6-digit passcode"
-        navigationItem.leftBarButtonItem = UIBarButtonItem(
-            barButtonSystemItem: .close, target: self, action: #selector(didTapCloseButton))
-        button.isHidden = true
-    }
-
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        Tracker.trackEvent(.changePasscodeEnterNew)
-    }
-
-    @objc func didTapCloseButton() {
-        completion()
-    }
-
-    override func willChangeText(_ text: String) {
-        super.willChangeText(text)
-        if text.count == passcodeLength {
-            let vc = RepeatChangedPasscodeViewController(passcode: text, completionHandler: completion)
-            navigationController?.pushViewController(vc, animated: true)
-        }
-    }
-}
-
-class RepeatChangedPasscodeViewController: PasscodeViewController {
-    var passcode: String!
-
-    convenience init(passcode: String, completionHandler: @escaping () -> Void) {
-        self.init(namedClass: PasscodeViewController.self)
-        self.passcode = passcode
-        completion = completionHandler
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        navigationItem.title = "Change Passcode"
-        promptLabel.text = "Repeat the 6-digit passcode"
-        button.isHidden = true
-    }
-
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        Tracker.trackEvent(.changePasscodeRepeat)
-    }
-
-    override func willChangeText(_ text: String) {
-        super.willChangeText(text)
-        errorLabel.isHidden = true
-        if text == passcode {
-            do {
-                try App.shared.auth.changePasscode(newPasscodeInPlaintext: text)
-                App.shared.snackbar.show(message: "Passcode changed")
-                navigationController?.dismiss(animated: true) { [unowned self] in
-                    self.completion()
-                }
-            } catch {
-                showGenericError(description: "Failed to change passcode", error: error)
-            }
-        } else if text.count == passcodeLength {
-            showError("Passcodes don't match")
-        }
-    }
-}
